+++ Example/SnabblePayExample/AppDelegate.swift
@@ -1,0 +1,15 @@
+//
+//  AppDelegate.swift
+//  SnabblePayExample
+//
+//  Created by Andreas Osberghaus on 2023-01-16.
+//
+
+import Foundation
+import UIKit
+
+class AppDelegate: NSObject, UIApplicationDelegate {
+    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]? = nil) -> Bool {
+        return true
+    }
+}