+++ Example/SnabblePayExample/SnabblePayExampleApp.swift
@@ -11,6 +11,8 @@ import SnabblePayNetwork
 
 @main
 struct SnabblePayExampleApp: SwiftUI.App {
+    @UIApplicationDelegateAdaptor(AppDelegate.self) var appDelegate
+
     let networkConfig: NetworkConfig = .init(
         customUrlScheme: "snabble-pay",
         apiKey: "IO2wX69CsqZUQ3HshOnRkO4y5Gy/kRar6Fnvkp94piA2ivUun7TC7MjukrgUKlu7g8W8/enVsPDT7Kvq28ycw=="