+++ Sources/Network/Core/Authenticator.swift
@@ -107,4 +107,9 @@ class Authenticator {
             return publisher
         }
     }
+
+    func reset() {
+        token = nil
+        app = nil
+    }
 }